Remove Message_p file - No need to hide private class from the public header
