+++ Shared/messages/Message.h
@@ -107,6 +107,23 @@ private:
   QSharedDataPointer<MessageData> d;
 };
 
+class MessageData : public QSharedData
+{
+public:
+  MessageData();
+  MessageData(Message::MessageAction messageAction, const Esri::ArcGISRuntime::Geometry& geometry);
+  MessageData(const MessageData& other);
+  ~MessageData();
+
+  Message::MessageAction messageAction = Message::MessageAction::Unknown;
+  QVariantMap attributes;
+  Esri::ArcGISRuntime::Geometry geometry;
+  QString messageId;
+  QString messageName;
+  QString messageType;
+  QString symbolId;
+};
+
 } // Dsa
 
 Q_DECLARE_METATYPE(Dsa::Message::MessageAction)
